+++ core/src/Language/Mimsa/Core/Types/AST/DataType.hs
@@ -51,7 +51,7 @@ renderDataType :: DataType -> Doc style
 renderDataType (DataType tyCon vars' constructors') =
   "type"
     <+> prettyDoc tyCon
-    <> printVars vars'
+      <> printVars vars'
     <+> if M.null constructors'
       then mempty
       else
