+++ core/src/Language/Mimsa/Core/Types/Module/Module.hs
@@ -132,15 +132,15 @@ printPaired (MTFunction _ fn arg) (MyLambda _ ident body) =
     <> prettyDoc ident
     <+> ":"
     <+> prettyDoc fn
-    <> ")"
-    <> line
-    <> printPaired arg body
+      <> ")"
+      <> line
+      <> printPaired arg body
 printPaired mt expr =
   ":"
     <+> prettyDoc mt
     <+> "="
-    <> line
-    <> indentMulti 2 (prettyDoc expr)
+      <> line
+      <> indentMulti 2 (prettyDoc expr)
 
 printDefinition :: Module ann -> DefIdentifier -> Expr Name ann -> Doc a
 printDefinition mod' def expr =
@@ -153,14 +153,14 @@ printDefinition mod' def expr =
           (MyAnnotation _ mt rest) ->
             "def"
               <+> prettyDoc name
-              <> line
-              <> indentMulti 2 (printPaired mt rest)
+                <> line
+                <> indentMulti 2 (printPaired mt rest)
           other ->
             "def"
               <+> prettyDoc name
               <+> "="
-              <> line
-              <> indentMulti 2 (prettyDoc other)
+                <> line
+                <> indentMulti 2 (prettyDoc other)
         DIInfix infixOp ->
           "infix" <+> prettyDoc infixOp <+> "=" <+> prettyDoc expr
         DIType _ -> error "printDefinition is printing type oh no"
