+++ smol-core/src/Smol/Core/Typecheck/Types/TCError.hs
@@ -1,0 +1,31 @@
+{-# LANGUAGE DeriveFoldable #-}
+{-# LANGUAGE DerivingStrategies #-}
+
+module Smol.Core.Typecheck.Types.TCError
+  ( TCError(..),
+  )
+where
+
+import Data.Set (Set)
+import Smol.Core.Types
+import Smol.Core.Types.PatternMatchError (PatternMatchError)
+
+data TCError ann
+  = TCUnknownError
+  | TCCouldNotFindVar ann (ResolvedDep Identifier)
+  | TCCouldNotFindGlobal Identifier (Set Identifier)
+  | TCTypeMismatch (ResolvedType ann) (ResolvedType ann)
+  | TCTupleSizeMismatch Int (ResolvedType ann)
+  | TCExpectedTuple (ResolvedType ann)
+  | TCExpectedFunction (ResolvedType ann)
+  | TCRecordMissingItems (Set Identifier)
+  | TCExpectedRecord (ResolvedType ann)
+  | TCInfixMismatch Op (ResolvedType ann) (ResolvedType ann)
+  | TCPatternMismatch (Pattern ResolvedDep ann) (ResolvedType ann)
+  | TCUnknownConstructor (ResolvedDep Constructor) [Constructor]
+  | TCConstructorArgumentMismatch (ResolvedDep Constructor) Int Int -- expected, actual
+  | TCExpectedConstructorType (ResolvedType ann)
+  | TCCompoundTypeInEquality (ResolvedType ann) -- for now we only do primitive equality
+  | TCPatternMatchError (PatternMatchError (ResolvedType ann))
+  deriving stock (Eq, Ord, Show, Foldable)
+
